Отрисовка случайно созданной вселенной

+++ GUI/Views/BodyWindow.xaml.cs
@@ -39,7 +39,7 @@ namespace GUI.Views
             {
                 double x = double.Parse(tbX.Text);
                 double y = double.Parse(tbY.Text);
-                byte d = byte.Parse(tbD.Text);
+                uint d = uint.Parse(tbD.Text);
                 uint mass = uint.Parse(tbMass.Text);
                 int vx = int.Parse(tbVx.Text);
                 int vy = int.Parse(tbVy.Text);
